backend/requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
